rf: move `test_get_linkml_err_counts()`
